Add support for translation notes
To add a translation note, use it as a filter. For example,
consider
   {% Free %}
If we want to say that this means free as in no monetary value,
wrap it around in {{ }} and apply the tl_note filter:
   {% {{ "Free" | tl_note: Free as in no monetary value }} %}
The tl_note parameter will be ignored when this Liquid filter is
parsed, but the entirety will be used as the keyed string. Just
make sure that translators don't remove the quotes around free.